core/requirements.txt
certifi==2018.11.29
Django==2.1.3
django-admin-sortable2==0.6.21
django-ckeditor==5.6.1
django-environ==0.4.5
django-js-asset==1.1.0
django-suit @ https://github.com/darklow/django-suit/tarball/v2
easy-thumbnails==2.5
idna==2.8
pyTelegramBotAPI==3.6.6
pytz==2018.7
requests==2.7.0
six==1.12.0
urllib3==1.24.1